@@ -4,22 +4,27 @@ import android.content.ActivityNotFoundException;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ListView;
+import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.crashlytics.android.Crashlytics;
 import com.crashlytics.android.answers.Answers;
 import com.crashlytics.android.answers.CustomEvent;
-import com.firebase.ui.database.FirebaseListAdapter;
-import com.firebase.ui.database.FirebaseListOptions;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.Query;
+import com.firebase.ui.firestore.paging.FirestorePagingAdapter;
+import com.firebase.ui.firestore.paging.FirestorePagingOptions;
+import com.google.firebase.firestore.CollectionReference;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.Query;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.paging.PagedList;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import hotchemi.android.rate.AppRate;
@@ -29,36 +34,29 @@ import me.cutmail.disasterapp.model.Entry;
 import timber.log.Timber;
 
 
-public class MainActivity extends AppCompatActivity implements AdapterView.OnItemClickListener {
+public class MainActivity extends AppCompatActivity {
 
-    @BindView(R.id.listView)
-    ListView listView;
+    @BindView(R.id.paging_recycler)
+    RecyclerView recyclerView;
 
-    private FirebaseListAdapter<Entry> adapter;
+    private FirebaseFirestore firestore;
+    private CollectionReference collectionReference;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         ButterKnife.bind(this);
+
+        firestore = FirebaseFirestore.getInstance();
+        collectionReference = firestore.collection("entries");
+
         setupLayout();
 
         Fabric.with(this, new Crashlytics());
         setupRateDialog();
 
         AppRate.showRateDialogIfMeetsConditions(this);
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        adapter.startListening();
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        adapter.stopListening();
     }
 
     private void setupRateDialog() {
@@ -89,21 +87,47 @@ public class MainActivity extends AppCompatActivity implements AdapterView.OnIte
     }
 
     private void setupLayout() {
-        Query query = FirebaseDatabase.getInstance()
-                .getReference("entries");
+        Query query = collectionReference.orderBy("title");
 
-        FirebaseListOptions<Entry> options = new FirebaseListOptions.Builder<Entry>()
-                .setQuery(query, Entry.class)
-                .setLayout(android.R.layout.simple_list_item_1)
+        PagedList.Config config = new PagedList.Config.Builder()
+                .setEnablePlaceholders(false)
+                .setPrefetchDistance(10)
+                .setPageSize(20)
                 .build();
 
-        adapter = new FirebaseListAdapter<Entry>(options) {
-            @Override protected void populateView(View v, Entry entry, int position) {
-                ((TextView) v.findViewById(android.R.id.text1)).setText(entry.getTitle());
+        FirestorePagingOptions<Entry> options = new FirestorePagingOptions.Builder<Entry>()
+                .setLifecycleOwner(this)
+                .setQuery(query, config, Entry.class)
+                .build();
+
+        FirestorePagingAdapter<Entry, ItemViewHolder> adapter = new FirestorePagingAdapter<Entry, ItemViewHolder>(options) {
+            @Override protected void onBindViewHolder(@NonNull ItemViewHolder holder, int position, @NonNull Entry entry) {
+                holder.bind(entry);
+            }
+
+            @NonNull @Override public ItemViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+                View view = LayoutInflater.from(parent.getContext())
+                        .inflate(android.R.layout.simple_list_item_1, parent, false);
+                return new ItemViewHolder(view);
             }
         };
-        listView.setAdapter(adapter);
-        listView.setOnItemClickListener(this);
+
+        recyclerView.setLayoutManager(new LinearLayoutManager(this));
+        recyclerView.setAdapter(adapter);
+
+//
+//        FirebaseListOptions<Entry> options = new FirebaseListOptions.Builder<Entry>()
+//                .setQuery(query, Entry.class)
+//                .setLayout(android.R.layout.simple_list_item_1)
+//                .build();
+//
+//        adapter = new FirebaseListAdapter<Entry>(options) {
+//            @Override protected void populateView(View v, Entry entry, int position) {
+//                ((TextView) v.findViewById(android.R.id.text1)).setText(entry.getTitle());
+//            }
+//        };
+//        listView.setAdapter(adapter);
+//        listView.setOnItemClickListener(this);
     }
 
     private void openAbout() {
@@ -137,12 +161,27 @@ public class MainActivity extends AppCompatActivity implements AdapterView.OnIte
         }
     }
 
-    @Override
-    public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-        if (adapter != null) {
-            Entry entry = adapter.getItem(position);
-            Intent intent = EntryDetailActivity.createIntent(this, entry.getTitle(), entry.getUrl());
-            startActivity(intent);
+//    @Override
+//    public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+//        if (adapter != null) {
+//            Entry entry = adapter.getItem(position);
+//            Intent intent = EntryDetailActivity.createIntent(this, entry.getTitle(), entry.getUrl());
+//            startActivity(intent);
+//        }
+//    }
+
+    public static class ItemViewHolder extends RecyclerView.ViewHolder {
+
+        @BindView(android.R.id.text1)
+        TextView mTextView;
+
+        ItemViewHolder(View itemView) {
+            super(itemView);
+            ButterKnife.bind(this, itemView);
+        }
+
+        void bind(Entry entry) {
+            mTextView.setText(entry.getTitle());
         }
     }
 }
